integrated functionality of _m2l subclass to superclass
