+++ hierarc/LensPosterior/kin_constraints_composite.py
@@ -16,8 +16,8 @@ class KinConstraintsComposite(KinConstraints):
         z_source,
         gamma_in_array,
         log_m2l_array,
-        rho0_array,
-        r_s_array,
+        kappa_s_array,
+        r_s_angle_array,
         theta_E,
         theta_E_error,
         gamma,
@@ -39,9 +39,9 @@ class KinConstraintsComposite(KinConstraints):
         num_psf_sampling=100,
         num_kin_sampling=1000,
         multi_observations=False,
-        kappa_s_array=None,
-        r_s_angle_array=None,
-        amp_2_luminosity=1,
+        rho0_array=None,
+        r_s_array=None,
+        is_m2l_population_level=True,
     ):
         """
 
@@ -86,9 +86,6 @@ class KinConstraintsComposite(KinConstraints):
             kwargs_seeing as lists of multiple observations
         :param kappa_s_array: array of generalized NFW profile's convergence normalization at the scale radius
         :param r_s_angle_array: array of halo scale radii in arcsecond
-        :param amp_2_luminosity: conversion from light profile amplitudes to luminosity
-            (e.g., considering user-preferred distance modulus, k-correction, zero-point, unit conversion, etc.);
-            should be so that amplitude * amp_2_luminosity = luminosity
         """
 
         if (
@@ -144,25 +141,32 @@ class KinConstraintsComposite(KinConstraints):
             multi_observations=multi_observations,
         )
 
-        if kappa_s_array is None and r_s_angle_array is None:
-            self._rho_s_array = rho0_array
-            self._r_scale_array = r_s_array
-            if len(rho0_array) != len(r_s_array):
-                raise ValueError("rho0 and r_s must have the same length!")
-
+        if self.arrays_check(kappa_s_array, r_s_angle_array):
+            self._kappa_s_array = kappa_s_array
+            self._r_scale_angle_array = r_s_angle_array
+        elif self.arrays_check(rho0_array, r_s_array):
             self._kappa_s_array, self._r_scale_angle_array = self.get_kappa_s_r_s_angle(
                 rho0_array, r_s_array
             )
         else:
-            if len(kappa_s_array) != len(r_s_angle_array):
-                raise ValueError("kappa_s and r_s_angle must have the same length!")
-
-            self._kappa_s_array = kappa_s_array
-            self._r_scale_angle_array = r_s_angle_array
+            raise ValueError(
+                "Both kappa_s_array and r_s_angle_array, or rho0_array and r_s_array must be arrays of the same length!"
+            )
 
         self.gamma_in_array = gamma_in_array
         self.log_m2l_array = log_m2l_array
-        self.amp_2_luminosity = amp_2_luminosity
+        self._is_m2l_population_level = is_m2l_population_level
+
+        if not is_m2l_population_level and not self.arrays_check(self._kappa_s_array, log_m2l_array):
+            raise ValueError(
+                "log_m2l_array must have the same length as rho0_array or kappa_s_array!"
+            )
+
+    @staticmethod
+    def arrays_check(array0, array1):
+        if array0 is None or array1 is None:
+            return False
+        return len(array0) == len(array1) and len(array0) > 0
 
     def get_kappa_s_r_s_angle(self, rho_s, r_scale):
         """Computes the surface mass density of the NFW halo at the scale radius.
@@ -183,12 +187,21 @@ class KinConstraintsComposite(KinConstraints):
             the mean values instead
         """
         if no_error is True:
-            return (
-                np.mean(self._kappa_s_array),
-                np.mean(self._r_scale_angle_array),
-                self._r_eff,
-                1,
-            )
+            if self._is_m2l_population_level:
+                return (
+                    np.mean(self._kappa_s_array),
+                    np.mean(self._r_scale_angle_array),
+                    self._r_eff,
+                    1,
+                )
+            else:
+                return (
+                    np.mean(self._kappa_s_array),
+                    np.mean(self._r_scale_angle_array),
+                    np.mean(self.log_m2l_array),
+                    self._r_eff,
+                    1,
+                )
 
         random_index = np.random.randint(low=0, high=len(self._kappa_s_array))
         kappa_s_draw = self._kappa_s_array[random_index]
@@ -200,7 +213,11 @@ class KinConstraintsComposite(KinConstraints):
         )
         r_eff_draw = delta_r_eff * self._r_eff
 
-        return kappa_s_draw, r_scale_angle_draw, r_eff_draw, delta_r_eff
+        if self._is_m2l_population_level:
+            return kappa_s_draw, r_scale_angle_draw, r_eff_draw, delta_r_eff
+        else:
+            log_m2l_draw = self.log_m2l_array[random_index]
+            return kappa_s_draw, r_scale_angle_draw, log_m2l_draw, r_eff_draw, delta_r_eff
 
     def model_marginalization(self, num_sample_model=20):
         """
@@ -218,12 +235,19 @@ class KinConstraintsComposite(KinConstraints):
             j_kin = np.nan
             c = 0
             while np.isnan(j_kin) and c < 10:
-                j_kin = self.j_kin_draw_composite(
-                    self.kwargs_anisotropy_base,
-                    np.mean(self.gamma_in_array),
-                    np.mean(self.log_m2l_array),
-                    no_error=False,
-                )
+                if self._is_m2l_population_level:
+                    j_kin = self.j_kin_draw_composite(
+                        self.kwargs_anisotropy_base,
+                        np.mean(self.gamma_in_array),
+                        np.mean(self.log_m2l_array),
+                        no_error=False,
+                    )
+                else:
+                    j_kin = self.j_kin_draw_composite_m2l(
+                        self.kwargs_anisotropy_base,
+                        np.mean(self.gamma_in_array),
+                        no_error=False,
+                    )
                 c += 1
             j_kin_matrix[i, :] = j_kin
 
@@ -250,7 +274,7 @@ class KinConstraintsComposite(KinConstraints):
         kwargs_lens_stars = copy.deepcopy(self._kwargs_lens_light[0])
 
         kwargs_lens_stars["amp"] *= (
-            10**log_m2l / self.lensCosmo.sigma_crit_angle * self.amp_2_luminosity
+            10**log_m2l / self.lensCosmo.sigma_crit_angle
         )
 
         kwargs_lens_stars["sigma"] *= delta_r_eff
@@ -271,6 +295,56 @@ class KinConstraintsComposite(KinConstraints):
         ]
 
         print(kwargs_lens)
+
+        j_kin = self.velocity_dispersion_map_dimension_less(
+            kwargs_lens=kwargs_lens,
+            kwargs_lens_light=kwargs_light,
+            kwargs_anisotropy=kwargs_anisotropy,
+            r_eff=r_eff_draw,
+            theta_E=self._theta_E,  # send this to avoid unnecessary recomputation
+            gamma=self._gamma,  # send this to avoid unnecessary recomputation
+        )
+        return j_kin
+
+    def j_kin_draw_composite_m2l(self, kwargs_anisotropy, gamma_in, no_error=False):
+        """Similar function to j_kin_draw_composite, but now drawing from log_m2l distribution.
+
+        :param kwargs_anisotropy: keyword argument of anisotropy setting
+        :param gamma_in: power-law slope of the mass model
+        :param no_error: bool, if True, does not render from the uncertainty but uses
+            the mean values instead
+        :return: dimensionless kinematic component J() Birrer et al. 2016, 2019
+        """
+        (
+            kappa_s_draw,
+            r_scale_angle_draw,
+            log_m2l_draw,
+            r_eff_draw,
+            delta_r_eff,
+        ) = self.draw_lens(no_error=no_error)
+
+        kwargs_lens_stars = copy.deepcopy(self._kwargs_lens_light[0])
+
+        kwargs_lens_stars["amp"] *= (
+            log_m2l_draw / self.lensCosmo.sigma_crit_angle
+        )
+
+        kwargs_lens_stars["sigma"] *= delta_r_eff
+
+        kwargs_light = copy.deepcopy(self._kwargs_lens_light)
+        for kwargs in kwargs_light:
+            kwargs["sigma"] *= delta_r_eff
+
+        kwargs_lens = [
+            {
+                "Rs": r_scale_angle_draw,
+                "gamma_in": gamma_in,
+                "kappa_s": kappa_s_draw,
+                "center_x": 0,
+                "center_y": 0,
+            },
+            kwargs_lens_stars,
+        ]
 
         j_kin = self.velocity_dispersion_map_dimension_less(
             kwargs_lens=kwargs_lens,
@@ -313,6 +387,9 @@ class KinConstraintsComposite(KinConstraints):
             "log_m2l_array": self.log_m2l_array,
             "param_scaling_grid_list": ani_scaling_grid_list,
         }
+
+        if not self._is_m2l_population_level:
+            kwargs_likelihood["log_m2l_array"] = None
         return kwargs_likelihood
 
     def anisotropy_scaling(self):
@@ -320,13 +397,21 @@ class KinConstraintsComposite(KinConstraints):
 
         :return: anisotropy scaling grid along the axes defined by ani_param_array
         """
-        j_ani_0 = self.j_kin_draw_composite(
-            self.kwargs_anisotropy_base,
-            np.mean(self.gamma_in_array),
-            np.mean(self.log_m2l_array),
-            no_error=True,
-        )
-        return self._anisotropy_scaling_relative(j_ani_0)
+        if self._is_m2l_population_level:
+            j_ani_0 = self.j_kin_draw_composite(
+                self.kwargs_anisotropy_base,
+                np.mean(self.gamma_in_array),
+                np.mean(self.log_m2l_array),
+                no_error=True,
+            )
+            return self._anisotropy_scaling_relative(j_ani_0)
+        else:
+            j_ani_0 = self.j_kin_draw_composite_m2l(
+                self.kwargs_anisotropy_base,
+                np.mean(self.gamma_in_array),
+                no_error=True,
+            )
+            return self._anisotropy_scaling_relative_m2l(j_ani_0)
 
     def _anisotropy_scaling_relative(self, j_ani_0):
         """Anisotropy scaling relative to a default J prediction.
@@ -385,6 +470,61 @@ class KinConstraintsComposite(KinConstraints):
                         )
                         for m, j_kin in enumerate(j_kin_ani):
                             ani_scaling_grid_list[m][i, k, l] = j_kin / j_ani_0[m]
+        else:
+            raise ValueError("anisotropy model %s not valid." % self._anisotropy_model)
+        return ani_scaling_grid_list
+
+    def _anisotropy_scaling_relative_m2l(self, j_ani_0):
+        """Similar function to _anisotropy_scaling_relative, but instead drawing log_m2l from a distribution.
+
+        :param j_ani_0: default J() prediction for default anisotropy
+        :return: list of arrays (for the number of measurements) according to anisotropy
+            scaling
+        """
+        num_data = len(self._sigma_v_measured)
+
+        if self._anisotropy_model == "GOM":
+            ani_scaling_grid_list = [
+                np.zeros(
+                    (
+                        len(self.ani_param_array[0]),
+                        len(self.ani_param_array[1]),
+                        len(self.gamma_in_array),
+                    )
+                )
+                for _ in range(num_data)
+            ]
+            for i, a_ani in enumerate(self.ani_param_array[0]):
+                for j, beta_inf in enumerate(self.ani_param_array[1]):
+                    for k, g_in in enumerate(self.gamma_in_array):
+                        kwargs_anisotropy = self.anisotropy_kwargs(
+                            a_ani=a_ani, beta_inf=beta_inf
+                        )
+                        j_kin_ani = self.j_kin_draw_composite_m2l(
+                            kwargs_anisotropy, g_in, no_error=True
+                        )
+
+                        for m, j_kin in enumerate(j_kin_ani):
+                            ani_scaling_grid_list[m][i, j, k] = j_kin / j_ani_0[m]
+                            # perhaps change the order
+        elif self._anisotropy_model in ["OM", "const"]:
+            ani_scaling_grid_list = [
+                np.zeros(
+                    (
+                        len(self.ani_param_array),
+                        len(self.gamma_in_array),
+                    )
+                )
+                for _ in range(num_data)
+            ]
+            for i, a_ani in enumerate(self.ani_param_array):
+                for k, g_in in enumerate(self.gamma_in_array):
+                    kwargs_anisotropy = self.anisotropy_kwargs(a_ani)
+                    j_kin_ani = self.j_kin_draw_composite_m2l(
+                        kwargs_anisotropy, g_in, no_error=True
+                    )
+                    for m, j_kin in enumerate(j_kin_ani):
+                        ani_scaling_grid_list[m][i, k] = j_kin / j_ani_0[m]
         else:
             raise ValueError("anisotropy model %s not valid." % self._anisotropy_model)
         return ani_scaling_grid_list
